Update DB schema to allow storing multi-extension FITS files.
Storing multi-ext FITS files is complicated because it's
not trivial to unravel them (leaving potentially many images
without trails) and it's not easy to store only the relevant
data. As was discussed, we store the whole file, and live
with a little bit of non-user-friendliness.

Fix the poor naming scheme of UploadProcessor attributes,
what was `upload` is `fileWrapper`, what is `upload` now
is the UploadInfo model.

Add UploadInfo model, which can be used to track what IP
uploaded which data and when.

Reset migration history.

Refactor the view and intiial processing code, punting
much of the functionality into the UploadProcessor.

Replace metadata_translator_name column by standardizer
and processor name columns. Add special naming convention
for astro_metadata_translator classes. Default the rest
to their class names.

Fix tests.