Add an example of how ModelForms work.
Had not yet had the time to install the 3rd party app
"django_tables2" and give it a good shake. The table class would
still live in the views.py and would give yet another way to map
our models to a table in this case, so I removed all that I
percieved as extraneous code.

Added verbose comments.

In short:

1) visiting home page and clicking query on the navigation
takes us to the root url of our query app. This does not by default
render any tables.
2) Submitting a query further redirects us to a query/results url
which by default always renders a table, with or without results.

How:
The ModelForm is used to map fields of our DB schema models onto
entry, text or date inut boxes.
The model form is rendered by the templating engine. That python
code becomes HTML.
We add an additional button to the form in the HTML. That button
activates an action, redirecting us to query/results url, by sending
a POST type request to our server.
We instantiate our model form again from the POST request and
validate input.
We access the posted data of our form, currently just the instrument,
and execute a query. We should probaby call clean data and access the
cleaned data. We should also probably define what clean method does
on our form (such as blocks all escape characters, converts everything
to lowercase letters etc..).
We return the queried results back to our templating engine, which
then renders them as HTML table.
In the case we somehow find ourselfs invoking print_results (by
visiting query/results URL) without making a POST request we render
the table header with no results and the form with no input.
We can rely on the fact we marked the instrument field as a required
field to skip checking whether there is any content in it. For example,
if you wanted to do the same for telescope, you would have to verify
the user did not just leave the box empty, and then issue an appropriate
query, one that would select * from...