Unify locations of default configuration.
Change the settings.py so that path configration occurs on top.
Separate the configuration of locations where we store data from
the static and media location used atm as a crutch.
Add a env var that allows us to point to a different config file than
the default one.
In combination with making all paths in the DB relative ones, this
affords us the opportunity to move our test data to a different repo.